Cleaning up the exporter.
